I have managed to make the program remove a template parameter and modify the using statement to use the old template parameter. I have also made the program output a new source file with the modified code. In essence the small test program I aimed for works at the moment.
However, I am not too pleased by the way I remove the template parameter i will ask the Clang mailing list if thier is an alternative method. Basically the rewriter ignores whitespace when you wish to remove it, so I ended up creating a work around that inserts a garbage character then removes it to get rid of the space. Not ideal.